starting to make phase be a float too

+++ dsp/variable_wave_oscillator.hpp
@@ -17,12 +17,10 @@ class VariWaveOsc {
 
         void init();
         void update_params(float wave);
-        int16_t process_sample(uint32_t phase);
+        int16_t process_sample(float phase);
     
     private:
-        // const int16_t* waves_[NUM_WAVES];
         const float* reso_waves_[NUM_RESO_WAVES];
-        // bool knob_left_;
         int32_t wave_val_i_;
         float wave_val_f_;
 };
